+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Master
-- ???
+- Required Forwardable on Collection to resolve NameError [\#44](https://github.com/RubyMoney/monetize/issues/44)
 
 ## 1.3.0
 - Add NilClass extension